requirements/base.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.2
inflection==0.5.1
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
uritemplate==4.1.1
virtualenv==20.27.1
python-dotenv==1.0.1
faker==33.0.0
sphinx==8.1.3
pytest==8.3.3
